requirements.txt
langchain
langchain-community
langchain-openai
pypdf
faiss-cpu
python-dotenv
tiktoken
yfinance
transformers
torch
accelerate
bitsandbytes
langchain-huggingface
torchvision
sentencepiece
accelerate>=0.26.0
bitsandbytes>=0.43.0